Add process_tweets.py and Docker setup
The process_tweets.py file will retrieve tweets from Kafka, classify
them, and then push the ones that appear to have hate speech in them to
Postgres. To use this, you need to have the `DB_` variables in your
environment or you need to add them to the `.env` file and use
docker-compose to run the script: `docker-compose up --build -d`.
Currently all the Kafka values are hard-coded, so it would probably be
good to make those environment variables as well at some point.