Make Dash app connect to Postgres DB
The app now uses the `python-dotenv` package to get sensitive
credentials from a .env file. Also connects to a Postgres DB to collect
tweets for processing and display. Includes a `server` variable for use
by gunicorn in the production deployment. Lastly, updates the
requirements.txt file to include the rest of the required packages.